app/requirements.txt
speedtest-cli==2.1.2
SQLAlchemy==1.3.22